+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 6.5.0
+
+- Add default period and delay managers.
+- Add support for getName in Worker that replaces the default name.
+
 ### 6.4.4
 
 - Fix Periodic throwing wrong delete exception.